Refactor profile and chat UI components
