+++ transcendental_resonance_frontend/ui/__init__.py
@@ -1,0 +1,10 @@
+"""UI helper exports."""
+from .profile_card import render_profile_card, inject_profile_styles, DEFAULT_USER
+from .chat_ui import render_chat_ui
+
+__all__ = [
+    "render_profile_card",
+    "inject_profile_styles",
+    "DEFAULT_USER",
+    "render_chat_ui",
+]
